+++ pytket/extensions/aqt/multi_zone_architecture/circuit/multizone_circuit.py
@@ -1,10 +1,24 @@
+# Copyright 2020-2023 Cambridge Quantum Computing
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import itertools
 import logging
 from copy import deepcopy
 from dataclasses import dataclass
 from dataclasses import field
 from enum import Enum
-from typing import Any, Optional
+from typing import Any, Optional, Iterator
 
 from sympy import symbols, Expr
 
@@ -17,9 +31,9 @@ from ..architecture import EdgeType
 from ..architecture import MultiZoneArchitecture
 from ..architecture import source_edge_type
 from ..architecture import target_edge_type
-from ..macro_architechture_graph import empty_macro_arch_from_backend
-from ..macro_architechture_graph import MultiZoneMacroArch
-from ..macro_architechture_graph import ZoneId
+from ..macro_architecture_graph import empty_macro_arch_from_architecture
+from ..macro_architecture_graph import MultiZoneMacroArch
+from ..macro_architecture_graph import ZoneId
 
 ParamType = Expr | float
 
@@ -194,11 +208,13 @@ class MultiZoneCircuit:
         **kwargs: str,
     ):
         self.architecture = multi_zone_arch
-        self.macro_arch = empty_macro_arch_from_backend(multi_zone_arch)
+        self.macro_arch = empty_macro_arch_from_architecture(multi_zone_arch)
         self.pytket_circuit = Circuit(*args, **kwargs)
         self.qubit_to_zones = {}
         self.initial_zone_to_qubits = initial_zone_to_qubits
-        self.zone_to_qubits = {zone.id: [] for zone in multi_zone_arch.zones}
+        self.zone_to_qubits = {
+            zone_id: [] for zone_id, _ in enumerate(multi_zone_arch.zones)
+        }
         self.multi_zone_operations = {
             qubit: [] for qubit in range(multi_zone_arch.n_qubits_max)
         }
@@ -215,6 +231,9 @@ class MultiZoneCircuit:
             init_def_circ = Circuit(len(qubit_list))
             custom_init = CustomGateDef("INIT", init_def_circ, [dz])
             self.pytket_circuit.add_custom_gate(custom_init, [zone], qubit_list)
+
+    def __iter__(self) -> Iterator:
+        return self.pytket_circuit.__iter__()
 
     @property
     def is_compiled(self) -> bool:
@@ -252,13 +271,27 @@ class MultiZoneCircuit:
         for qubit in qubits:
             self._place_qubit(zone, qubit)
 
-    def move_qubit(self, qubit: int, new_zone: int) -> None:
+    def move_qubit(self, qubit: int, new_zone: int, precompiled: bool = False) -> None:
         """Move a qubit from its current zone to new_zone
 
         Calculates the needs "PSWAP" and "SHUTTLE" operations to implement move.
         Adds custom gates to underlying Circuit to signify move and prevent optimisation
         through the move.
         Raises error is move is not possible
+
+        If precompiled=False, the needed "PSWAP" and "SHUTTLE" operations are added to
+        lists for each qubit and "MoveBarriers" are added to underlying pytket circuit.
+        The "MoveBarriers" serve as markers to add the physical operations to the
+        circuit after compilation
+
+        If precompiled=True (should not be used for manual routing), the underlying
+        circuit is assumed to already be compiled (but not yet routed). the "PSWAP"
+        and "SHUTTLE" operations will be added to the circuit directly.
+
+        :param qubit: qubit to move
+        :param new_zone: zone to move it too
+        :param precompiled: whether the underlying pytket circuit has already been
+         compiled (but not yet routed)
         """
         if qubit not in self.qubit_to_zones:
             raise QubitPlacementError("Cannot move qubit that was never placed")
@@ -316,8 +349,9 @@ class MultiZoneCircuit:
             else:
                 position_in_zone = VirtualZonePosition.VirtualLeft
 
-        self.pytket_circuit.add_custom_gate(move_gate, [new_zone], [qubit])
-        self.add_move_barrier()
+        if not precompiled:
+            self.pytket_circuit.add_custom_gate(move_gate, [new_zone], [qubit])
+            self.add_move_barrier()
         old_zone_qubits.remove(qubit)
         if position_in_zone is VirtualZonePosition.VirtualLeft:
             self.zone_to_qubits[new_zone].insert(0, qubit)
@@ -325,6 +359,9 @@ class MultiZoneCircuit:
             self.zone_to_qubits[new_zone].append(qubit)
         self.qubit_to_zones[qubit].append(new_zone)
         self.multi_zone_operations[qubit].append(move_operations)
+        if precompiled:
+            for multi_op in move_operations:
+                multi_op.append_to_circuit(self)
 
     def add_gate(
         self,
@@ -358,6 +395,32 @@ class MultiZoneCircuit:
         new_circuit.zone_to_qubits = deepcopy(self.zone_to_qubits)
         new_circuit.multi_zone_operations = deepcopy(self.multi_zone_operations)
         return new_circuit
+
+    def get_n_shuttles(self) -> int:
+        """
+        Get the number of shuttles used to route the circuit to the architecture
+        """
+        count = 0
+        for cmd in self.pytket_circuit.get_commands():
+            op = cmd.op
+            optype = op.type
+            op_string = f"{op}"
+            if optype == OpType.CustomGate and "SHUTTLE" in op_string:
+                count += 1
+        return count
+
+    def get_n_pswaps(self) -> int:
+        """
+        Get the number of pswaps used to route the circuit to the architecture
+        """
+        count = 0
+        for cmd in self.pytket_circuit.get_commands():
+            op = cmd.op
+            optype = op.type
+            op_string = f"{op}"
+            if optype == OpType.CustomGate and "PSWAP" in op_string:
+                count += 1
+        return count
 
     def validate(self) -> None:
         if self._is_compiled: